// plugins/cad/src/types/Function.ts
/**
 * Copyright 2022 Google LLC
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

export type Function = {
  apiVersion: string;
  kind: string;
  metadata: FunctionMetadata;
  spec: FunctionSpec;
};

export type FunctionMetadata = {
  name: string;
  namespace?: string;
  creationTimestamp?: string;
};

export type FunctionSpec = {
  description: string;
  documentationUrl?: string;
  functionTypes: string[];
  keywords: string[];
  image: string;
  repositoryRef: FunctionRepositoryRef;
};

export type FunctionRepositoryRef = {
  name: string;
};